Remove: ugly timeout implementation
Threading may hide exceptions when they come up.  When exceptions
happen, Observer should be triggered by crashing the entire autograder.
It is the responsibility of the custom test case implementer to
implement a workable timeout.

In the future, a wrapper that uses threading, checks for exceptions in
that thread, and kills the thread when timing out can work, but the
current implementation encourages sloppy timeout implementations and
hides errors when they come up, which is worse than not doing anything
and just letting it timeout/crash.